restructure endpoints
1. Merged /api/health into /api/status:
- removed exposing sensitive information about hardware status (should be done separately by monitoring software).
- removed hearthbeat-like endpoints — lightweight /api/status could be used instead

2. Split /api/signup:
- removed `POST /api/signup`: moved account creation to `POST /api/account` (public)
- removed `GET /api/signup/check/:pubkey` — duplicated by `GET /api/account/:pubkey` (public)
- the rest of /api/account endpoints are still auth protected (require NIP98 header)
- `GET /api/signup/pricing` → `GET /api/subscription/pricing`

3. Split `nip98auth` middleware in two parts:
- added `decodeNIP98header` to be used for ALL routes. It decodes NIP98 token if Authentication header is present, otherwise does nothing
- changed `nip98auth` to require `req.authenticatedPubkey` (set by `decodeNIP98header`). If `decodeNIP98header` wasn't used by mistake, `nip98auth` will call the same logic as a ssafeguard

This allows to make certain endpoints work differently for guests and authenticated users: e.g. they might return more data for authenticated user

4. Integration tests were split by endpoints instead of a single file
5. Integration tests now can run without Azure connection string (Azure Table communication is mocked by jest)
