+++ src/index.js
@@ -1,28 +1,4 @@
-// Early error handling setup before any other imports
-process.on('uncaughtException', (error) => {
-  console.error('💥 FATAL: Uncaught Exception during startup!');
-  console.error('Error:', error.message);
-  console.error('Stack:', error.stack);
-  console.error('This error occurred before the logger could be initialized.');
-  process.exit(1);
-});
-
-process.on('unhandledRejection', (reason, promise) => {
-  console.error('💥 FATAL: Unhandled Promise Rejection during startup!');
-  console.error('Reason:', reason);
-  console.error('Promise:', promise);
-  console.error('This error occurred before the logger could be initialized.');
-  process.exit(1);
-});
-
-// Load environment variables first
-try {
-  require('dotenv').config();
-  console.log('Environment variables loaded successfully');
-} catch (error) {
-  console.error('Failed to load environment variables:', error);
-  process.exit(1);
-}
+require('dotenv').config();
 
 // Core Node.js modules
 const express = require('express');
@@ -31,51 +7,24 @@ const helmet = require('helmet');
 const path = require('path');
 const fs = require('fs');
 const { accountsService } = require('./utils/AccountsTableService');
+const { apiKeyAuth, decodeNIP98Header } = require('./middleware/auth')
+const { errorHandler, notFoundHandler } = require('./middleware/errorHandler')
+const {
+  publicApiLimiter,
+  authenticatedLimiter,
+  adminLimiter,
+  paymentLimiter
+} = require('./middleware/rateLimiting')
+const logger = require('./utils/logger')
 
-// Import routes with error handling
-let subscriptionRoutes, notificationRoutes, statusRoutes, keyRoutes;
-let signupRoutes, adminRoutes, subscriptionManagementRoutes, accountRoutes;
+const subscriptionRoutes = require('./routes/subscription');
+const notificationRoutes = require('./routes/notification');
+const statusRoutes = require('./routes/status');
+const keyRoutes = require('./routes/key');
+const adminRoutes = require('./routes/admin');
+const subscriptionManagementRoutes = require('./routes/subscriptionManagement');
+const accountRoutes = require('./routes/account');
 
-try {
-  console.log('Loading route modules...');
-  subscriptionRoutes = require('./routes/subscription');
-  notificationRoutes = require('./routes/notification');
-  statusRoutes = require('./routes/status');
-  keyRoutes = require('./routes/key');
-  signupRoutes = require('./routes/signup');
-  adminRoutes = require('./routes/admin');
-  subscriptionManagementRoutes = require('./routes/subscriptionManagement');
-  accountRoutes = require('./routes/account');
-  console.log('Route modules loaded successfully');
-} catch (error) {
-  console.error('Failed to load route modules:', error);
-  console.error('Stack:', error.stack);
-  process.exit(1);
-}
-
-// Import middleware with error handling
-let apiKeyAuth, errorHandler, notFoundHandler;
-let publicApiLimiter, signupLimiter, authenticatedLimiter, adminLimiter, paymentLimiter;
-let logger;
-
-try {
-  console.log('Loading middleware modules...');
-  ({ apiKeyAuth } = require('./middleware/auth'));
-  ({ errorHandler, notFoundHandler } = require('./middleware/errorHandler'));
-  ({ 
-    publicApiLimiter, 
-    signupLimiter, 
-    authenticatedLimiter, 
-    adminLimiter, 
-    paymentLimiter 
-  } = require('./middleware/rateLimiting'));
-  logger = require('./utils/logger');
-  console.log('Middleware modules loaded successfully');
-} catch (error) {
-  console.error('Failed to load middleware modules:', error);
-  console.error('Stack:', error.stack);
-  process.exit(1);
-}
 
 // Startup validation and initialization
 async function validateStartupEnvironment() {
@@ -131,10 +80,10 @@ async function validateStartupEnvironment() {
     }    // Test Azure Table Storage initialization
     try {
       logger.info('Testing Azure Table Storage connection...');
-      
+
       // Test a simple operation on the accounts table      
       logger.info('Testing connection to accounts table...');
-      const testIterator = accountsService.tableClient.listEntities({ 
+      const testIterator = accountsService.tableClient.listEntities({
         queryOptions: { filter: "PartitionKey ne ''", top: 1 }
       });
       await testIterator.next();
@@ -161,10 +110,10 @@ async function validateStartupEnvironment() {
 
   } catch (error) {
     logger.error('Startup validation failed with unexpected error:', error);
-    return { 
-      success: false, 
-      errors: [`Startup validation failed: ${error.message}`], 
-      warnings 
+    return {
+      success: false,
+      errors: [`Startup validation failed: ${error.message}`],
+      warnings
     };
   }
 }
@@ -189,7 +138,7 @@ function initializeExpressApp() {
   try {
     const app = express();
     const PORT = process.env.PORT || 3000;
-    
+
     // Configure trust proxy based on environment
     if (process.env.NODE_ENV === 'test') {
       // In test environment, don't trust proxy to avoid rate limiting warnings
@@ -207,14 +156,14 @@ function initializeExpressApp() {
 
     // Static files
     app.use(express.static(path.join(__dirname, 'public')));    // Routes
+    app.use(decodeNIP98Header); // decode NIP98 Auth
     app.use('/api/subscription', publicApiLimiter, subscriptionRoutes);
     app.use('/api/notification', authenticatedLimiter, notificationRoutes); // Protected route
     app.use('/api/status', publicApiLimiter, statusRoutes);
     app.use('/api/key', publicApiLimiter, keyRoutes);
-    app.use('/api/signup', signupLimiter, signupRoutes); // Public signup route with strict limiting
     app.use('/api/admin', adminLimiter, adminRoutes); // Protected admin routes
     app.use('/api/subscription-management', paymentLimiter, subscriptionManagementRoutes); // Protected subscription management
-    app.use('/api/account', authenticatedLimiter, accountRoutes); // Protected account management
+    app.use('/api/account', accountRoutes); // Account management
 
     // Error handling middleware
     app.use(notFoundHandler);
@@ -231,13 +180,13 @@ function initializeExpressApp() {
 function setupGracefulShutdown(server) {
   const gracefulShutdown = (signal) => {
     logger.info(`Received ${signal}. Starting graceful shutdown...`);
-    
+
     server.close((err) => {
       if (err) {
         logger.error('Error during server shutdown:', err);
         process.exit(1);
       }
-      
+
       logger.info('Server closed successfully');
       process.exit(0);
     });
@@ -269,12 +218,12 @@ async function startServer() {
     if (!validation.success) {
       logger.error('Startup validation failed with the following errors:');
       validation.errors.forEach(error => logger.error(`  - ${error}`));
-      
+
       if (validation.warnings.length > 0) {
         logger.warn('Additional warnings:');
         validation.warnings.forEach(warning => logger.warn(`  - ${warning}`));
       }
-      
+
       logger.error('Cannot start server due to validation errors. Please fix the above issues and try again.');
       process.exit(1);
     }
@@ -289,11 +238,11 @@ async function startServer() {
       logger.info(`📍 Server running on port ${PORT}`);
       logger.info(`🌍 Environment: ${process.env.NODE_ENV || 'development'}`);
       logger.info(`⏰ Started at: ${new Date().toISOString()}`);
-      
+
       if (validation.warnings.length > 0) {
         logger.warn(`⚠️  ${validation.warnings.length} warning(s) detected during startup`);
       }
-      
+
       logger.info('='.repeat(50));
     });
 
@@ -328,7 +277,7 @@ process.on('uncaughtException', (error) => {
   logFunction('Error:', error.message);
   logFunction('Stack:', error.stack);
   logFunction('This is a fatal error. The process will exit.');
-  
+
   // Give the logger time to write before exiting
   setTimeout(() => {
     process.exit(1);
@@ -341,7 +290,7 @@ process.on('unhandledRejection', (reason, promise) => {
   logFunction('Reason:', reason);
   logFunction('Promise:', promise);
   logFunction('This indicates a programming error. Please review the code.');
-  
+
   // Don't exit immediately for unhandled rejections, just log
   // The process can continue but this should be investigated
 });
