+++ src/routes/account.js
@@ -6,18 +6,120 @@ const { paymentsService } = require('../utils/PaymentsTableService');
 const logger = require('../utils/logger');
 const { nip98Auth } = require('../middleware/auth');
 const logUserActivity = require('../middleware/logUserActivity');
+const {
+  signupLimiter,
+  authenticatedLimiter,
+} = require('../middleware/rateLimiting')
 
-// All account routes require NIP-98 authentication
-router.use(nip98Auth);
+// combined middleware to be used for routes requiring
+// authenticated user
+const authUser = [authenticatedLimiter, nip98Auth, logUserActivity]
+const publicApi = [signupLimiter]
 
-// Middleware to log user operations
-router.use(logUserActivity);
+/**
+ * Public endpoint for new user signups
+ * @route POST /api/account
+ */
+router.post('/', publicApi, async (req, res) => {
+  try {
+    const { pubkey, email, referralCode } = req.body;
+
+    if (!pubkey) {
+      return res.status(400).json({ error: 'Public key is required' });
+    }
+
+    // Check if user already exists
+    const existingUser = await accountsService.getEntity(pubkey, 'profile');
+    if (existingUser) {
+      return res.status(409).json({ error: 'User already exists' });
+    }
+
+    // Create user profile
+    const userProfile = {
+      pubkey,
+      email: email || null,
+      referralCode: referralCode || null,
+      signupDate: new Date().toISOString(),
+      status: 'active',
+      tier: 'free'
+    };
+
+    await accountsService.upsertEntity(pubkey, 'profile', userProfile);
+
+    // Initialize free subscription
+    const freeSubscription = {
+      tier: 'free',
+      isActive: true,
+      startDate: new Date().toISOString(),
+      expiryDate: null, // Free tier doesn't expire
+      billingCycle: null,
+      autoRenew: false,
+      dailyNotificationLimit: parseInt(process.env.FREE_TIER_DAILY_LIMIT) || 5
+    };
+
+    await subscriptionsService.upsertSubscription(pubkey, freeSubscription);
+
+    logger.info(`New user signup: ${pubkey.substring(0, 16)}... with email: ${email || 'none'}`);
+
+    res.status(201).json({
+      success: true,
+      message: 'User registered successfully',
+      user: {
+        pubkey,
+        tier: 'free',
+        signupDate: userProfile.signupDate
+      }
+    });
+  } catch (error) {
+    logger.error(`Error during signup: ${error.message}`);
+    res.status(500).json({ error: 'Failed to register user' });
+  }
+});
+
+/**
+ * Public endpoint to get user profile
+ * @route GET /api/account/:pubkey
+ */
+router.get('/:pubkey', publicApi, async (req, res) => {
+  try {
+    const targetPubkey = req.params.pubkey;
+
+    if (!targetPubkey) {
+      return res.status(400).json({ error: 'Public key is required' });
+    }
+
+    // Get user profile
+    const profile = await accountsService.getEntity(targetPubkey, 'profile');
+    if (!profile) {
+      return res.status(404).json({ error: 'User not found' });
+    }
+
+    // Get subscription status
+    const subscriptionStatus = await subscriptionsService.getSubscriptionStatus(targetPubkey);
+
+    // Public profile information
+    const publicProfile = {
+      pubkey: profile.pubkey,
+      signupDate: profile.signupDate,
+      tier: subscriptionStatus.tier,
+      isActive: subscriptionStatus.isActive
+    };
+
+    res.status(200).json({
+      success: true,
+      profile: publicProfile
+    });
+  } catch (error) {
+    logger.error(`Error getting user profile: ${error.message}`);
+    res.status(500).json({ error: 'Failed to get user profile' });
+  }
+});
 
 /**
  * Get own account information
- * @route GET /api/account/profile
+ * @route GET /api/account
  */
-router.get('/profile', async (req, res) => {
+router.get('/', authUser, async (req, res) => {
   try {
     const pubkey = req.authenticatedPubkey;
 
@@ -58,9 +160,9 @@ router.get('/profile', async (req, res) => {
 
 /**
  * Update account profile
- * @route PUT /api/account/profile
+ * @route PUT /api/account
  */
-router.put('/profile', async (req, res) => {
+router.put('/', authUser, async (req, res) => {
   try {
     const pubkey = req.authenticatedPubkey;
     const { email, metadata } = req.body;
@@ -102,7 +204,7 @@ router.put('/profile', async (req, res) => {
  * Get subscription details
  * @route GET /api/account/subscription
  */
-router.get('/subscription', async (req, res) => {
+router.get('/subscription', authUser, async (req, res) => {
   try {
     const pubkey = req.authenticatedPubkey;
 
@@ -131,7 +233,7 @@ router.get('/subscription', async (req, res) => {
  * Get payment history
  * @route GET /api/account/payments
  */
-router.get('/payments', async (req, res) => {
+router.get('/payments', authUser, async (req, res) => {
   try {
     const pubkey = req.authenticatedPubkey;
     const limit = parseInt(req.query.limit) || 25;
@@ -158,7 +260,7 @@ router.get('/payments', async (req, res) => {
  * Update subscription preferences
  * @route PUT /api/account/subscription/preferences
  */
-router.put('/subscription/preferences', async (req, res) => {
+router.put('/subscription/preferences', authUser, async (req, res) => {
   try {
     const pubkey = req.authenticatedPubkey;
     const { autoRenew, billingCycle } = req.body;
@@ -204,32 +306,32 @@ router.put('/subscription/preferences', async (req, res) => {
  * Get notification usage statistics
  * @route GET /api/account/usage
  */
-router.get('/usage', async (req, res) => {
+router.get('/usage', authUser, async (req, res) => {
   try {
     const pubkey = req.authenticatedPubkey;
 
     // Get current subscription to check limits
     const subscriptionStatus = await subscriptionsService.getSubscriptionStatus(pubkey);
-    
+
     // Get 24-hour notification count
     const count24h = await subscriptionsService.get24HourNotificationCount(pubkey);
-    
+
     // Get profile for total notifications
     const profile = await accountsService.getEntity(pubkey, 'profile');
 
     const usage = {
       current24Hours: count24h,
-      dailyLimit: subscriptionStatus.tier === 'free' 
+      dailyLimit: subscriptionStatus.tier === 'free'
         ? parseInt(process.env.FREE_TIER_DAILY_LIMIT) || 5
         : subscriptionStatus.tier === 'premium'
-        ? parseInt(process.env.PREMIUM_TIER_DAILY_LIMIT) || 50
-        : parseInt(process.env.PREMIUM_PLUS_TIER_DAILY_LIMIT) || 500,
+          ? parseInt(process.env.PREMIUM_TIER_DAILY_LIMIT) || 50
+          : parseInt(process.env.PREMIUM_PLUS_TIER_DAILY_LIMIT) || 500,
       totalAllTime: profile?.totalNotificationsSent || 0,
-      percentageUsed: Math.round((count24h / (subscriptionStatus.tier === 'free' 
+      percentageUsed: Math.round((count24h / (subscriptionStatus.tier === 'free'
         ? parseInt(process.env.FREE_TIER_DAILY_LIMIT) || 5
         : subscriptionStatus.tier === 'premium'
-        ? parseInt(process.env.PREMIUM_TIER_DAILY_LIMIT) || 50
-        : parseInt(process.env.PREMIUM_PLUS_TIER_DAILY_LIMIT) || 500)) * 100),
+          ? parseInt(process.env.PREMIUM_TIER_DAILY_LIMIT) || 50
+          : parseInt(process.env.PREMIUM_PLUS_TIER_DAILY_LIMIT) || 500)) * 100),
       tier: subscriptionStatus.tier
     };
 
@@ -248,7 +350,7 @@ router.get('/usage', async (req, res) => {
  * Delete account (self-deletion)
  * @route DELETE /api/account
  */
-router.delete('/', async (req, res) => {
+router.delete('/', authUser, async (req, res) => {
   try {
     const pubkey = req.authenticatedPubkey;
     const { confirmPubkey } = req.body;
@@ -263,7 +365,7 @@ router.delete('/', async (req, res) => {
     // 2. Process any pending refunds
     // 3. Delete user data according to GDPR/privacy requirements
     // 4. Send confirmation email
-    
+
     // For now, we'll just mark the profile as deleted
     const profile = await accountsService.getEntity(pubkey, 'profile');
     if (profile) {
@@ -274,7 +376,7 @@ router.delete('/', async (req, res) => {
         email: null, // Remove PII
         metadata: {} // Clear metadata
       };
-      
+
       await accountsService.upsertEntity(pubkey, 'profile', deletedProfile);
     }
 
@@ -287,7 +389,7 @@ router.delete('/', async (req, res) => {
         cancelledAt: new Date().toISOString(),
         autoRenew: false
       };
-      
+
       await subscriptionsService.upsertSubscription(pubkey, cancelledSubscription);
     }
