+++ src/routes/status.js
@@ -1,7 +1,11 @@
 const express = require('express');
 const router = express.Router();
 const os = require('os');
+const { adminAuth } = require('../middleware/auth');
 const logger = require('../utils/logger');
+const { accountsService } = require('../utils/AccountsTableService');
+const { subscriptionsService } = require('../utils/SubscriptionsTableService');
+const { paymentsService } = require('../utils/PaymentsTableService');
 
 /**
  * Get service status
@@ -12,22 +16,10 @@ router.get('/', async (req, res) => {
     const status = {
       service: 'Nostria Service',
       version: process.env.npm_package_version || '1.0.0',
-      uptime: process.uptime(),
       environment: process.env.NODE_ENV || 'development',
       timestamp: new Date().toISOString(),
-
-      // TODO: This information will not be provided in the future. This can be abused to validate if potential
-      // attacks is successful (increased memory usage, etc.).
-      system: {
-        platform: os.platform(),
-        arch: os.arch(),
-        memory: {
-          total: Math.round(os.totalmem() / (1024 * 1024)) + ' MB',
-          free: Math.round(os.freemem() / (1024 * 1024)) + ' MB',
-        }
-      }
     };
-    
+
     res.status(200).json(status);
   } catch (error) {
     logger.error(`Error getting service status: ${error.message}`);
@@ -36,11 +28,31 @@ router.get('/', async (req, res) => {
 });
 
 /**
- * Health check endpoint
- * @route GET /api/status/health
+ * Get signup statistics (public endpoint)
+ * @route GET /api/status/stats
  */
-router.get('/health', (req, res) => {
-  res.status(200).json({ status: 'ok' });
+router.get('/stats', async (req, res) => {
+  try {
+    // This is a simplified version - in production you might want to cache these stats
+    // and update them periodically rather than querying in real-time
+    
+    const stats = {
+      totalUsers: 'Coming soon', // Would require scanning all entities or maintaining counters
+      freeUsers: 'Coming soon',
+      premiumUsers: 'Coming soon',
+      premiumPlusUsers: 'Coming soon',
+      signupsToday: 'Coming soon',
+      message: 'Detailed statistics will be available in a future update'
+    };
+
+    res.status(200).json({
+      success: true,
+      stats
+    });
+  } catch (error) {
+    logger.error(`Error getting signup stats: ${error.message}`);
+    res.status(500).json({ error: 'Failed to get signup statistics' });
+  }
 });
 
 module.exports = router;
